Create Chef_and_Asteroids.py
Problem Statement 
-----------------------------

Chef enjoys playing Among Us with the other Chefs in the kitchen. However, he struggles with the asteroids task.

The task consists of N asteroids where the coordinates of the ith asteroid are (Xi,Yi).
In one second, each non-destroyed asteroid moves to the left by 1 unit.
Formally, if an asteroid's current position is (X,Y), then, after 11 second, it moves to the position (X−1,Y).
Note that, an asteroid can move to a negative X coordinate after some time.

The optimal strategy to play the game is to choose a non-destroyed asteroid having minimum X coordinate, and destroy it.
If multiple asteroids have minimum X coordinate, any one of them can be destroyed first.

However, Chef likes to take his chances. Each second, he chooses a non-destroyed asteroid, uniformly at random and destroys it.

Find the probability that Chef's strategy of randomly destroying the asteroids is same as the optimal strategy.

The probability can be written in the form of PQQP​ such that Q is not divisible by 1000000007. Output the value of P⋅Q−1 mod 1000000007.